Hi @ianlopshire, I make a pull request for supporting marshal Unicode.
Here is the benchmark:

```
benchmark                               old ns/op     new ns/op     delta
BenchmarkMarshal_MixedData_1-4          4135          3843          -7.06%
BenchmarkMarshal_MixedData_1000-4       3827841       3563933       -6.89%
BenchmarkMarshal_MixedData_100000-4     336806933     373349191     +10.85%
BenchmarkMarshal_String-4               1308          1069          -18.27%
BenchmarkMarshal_StringPtr-4            1266          1154          -8.85%
BenchmarkMarshal_Int64-4                1038          1059          +2.02%
BenchmarkMarshal_Float64-4              1381          1406          +1.81%

benchmark                               old allocs     new allocs     delta
BenchmarkMarshal_MixedData_1-4          48             48             +0.00%
BenchmarkMarshal_MixedData_1000-4       44009          44009          +0.00%
BenchmarkMarshal_MixedData_100000-4     4400015        4400015        +0.00%
BenchmarkMarshal_String-4               9              9              +0.00%
BenchmarkMarshal_StringPtr-4            9              9              +0.00%
BenchmarkMarshal_Int64-4                9              9              +0.00%
BenchmarkMarshal_Float64-4              12             12             +0.00%

benchmark                               old bytes     new bytes     delta
BenchmarkMarshal_MixedData_1-4          5248          5680          +8.23%
BenchmarkMarshal_MixedData_1000-4       1300042       1732044       +33.23%
BenchmarkMarshal_MixedData_100000-4     112738222     155938227     +38.32%
BenchmarkMarshal_String-4               4344          4376          +0.74%
BenchmarkMarshal_StringPtr-4            4344          4376          +0.74%
BenchmarkMarshal_Int64-4                4336          4368          +0.74%
BenchmarkMarshal_Float64-4              4400          4440          +0.91%
```

Please let me know if I missed any case or something that can be improved.

Regards,
Huy Dang