@ianlopshire 2019-07-11
@huydang284,

Thanks for submitting this! I apologize for the delay on getting it reviewed.

A couple of comments:

1. The decoder implementation has the Unicode support disabled by default. I think the encoder implementation needs to do the same.

2. I’m not sure if I agree with where the actually utf8 logic is placed. I think the different vale encoders could be updated to handle Unicode characters. Perhaps the `rawValue` type could be reused to accomplish this.

Let me know your thoughts.