@huydang284 2019-07-12
Hi @ianlopshire,

For your comments:
1. I added `useCodepointIndices` same as the decoder.
2. I'm not sure about using `rawValue` here. Could you please explain more?