@ianlopshire 2019-07-12
@huydang284,

The `rawValue` type is a way to store a value in a `[]byte` and lazily keep track of code point indices when necessary.

To utilize this, the signature of `valueEncoder` will need to change to `func(v reflect.Value, useCodepointIndices bool) (rawValue, error)` This will allow the individual value encoders to specify if the value they return needs to handle unicode codepoints. For example:

```go
func intEncoder(v reflect.Value, useCodepointIndices bool) (rawValue, error) {
	b := []byte(strconv.Itoa(int(v.Int())))
	// an int can only be ASCII so we pass false for useCodepointIndices.
	return newRawValue(b, false)
}

func stringEncoder(v reflect.Value, useCodepointIndices bool) (rawValue, error) {
	b := []byte(v.String())
	// a string can contain utf8 code points so we should the given
	// useCodepointIndices value.
	return newRawValue(b, useCodepointIndices)
}
```

The `structEncoder` will need to be updated with logic on how to combine the `rawValue`s into a single line. It may make sense to move the logic for that to a method on `rawValue` itself. Something like:

```go
func (r *rawValue) WriteAt(v rawValue, startPos, endPos int) error {
	// implement logic to write to the rawValue while respecting codepoint
	// indices.
}
```

The struct `structEncoder` implementation would then need to be updated to something along the lines of:

```go
func structEncoder(v reflect.Value, useCodepointIndices bool) (rawValue, error) {
	ss := cachedStructSpec(v.Type())

	// b only needs to be the length of the line. `rawValue.WriteAt` should take
	// care of expanding the length as needed.
        //
        // It may make sense to write a rawValue repeat function that automatically
	// gives the rawValue's internal []byte a capacity of len*UTFMax.
	b := bytes.Repeat([]byte(" "), ss.ll)
	dst, err := newRawValue(b, false)
	if err != nil {
		return rawValue{}, err
	}
	
	for i, spec := range ss.fieldSpecs {
		if !spec.ok {
			continue
		}

		val, err := newValueEncoder(v.Field(i).Type())(v.Field(i), useCodepointIndices bool)
		if err != nil {
			return rawValue{}, err
		}
		err = dst.WriteAt(val, spec.startPos-1, spec.endPos)
		if err != nil {
			return rawValue{}, err
		}
	}
	return dst, nil
}
```

Hopefully that helps! Let me know if you have additional questions.