@ianlopshire 2019-07-12
@Alex, Would you mind taking a look at the suggestion above. Any thoughts or suggestions? You first came up with the `rawValue` type.