@alex 2019-07-12
Off the cuff, it looks like this PR changes the code in roughly the places I'd expect. (I didn't look super closely if the implementation was correct though)

I don't think you need the full datastructure of `rawValue`, you don't need the codepoint indices, you just need to know how many codepoints are in the string if it's less than the limit (so you can pad appropriately), or where the cutoff point is if it's too long.