@ianlopshire 2019-07-12
@alex @huydang284,

The current implementation fails when the indices are not defined in order on the struct. For example: 

```go
type people struct {
	ID        int     `fixed:"1,5"`
	LastName  string  `fixed:"16,25"`
	FirstName string  `fixed:"6,15"`
	Grade     float64 `fixed:"26,30"`
}
```

This could be solved by making sure values are written in order, but I think that approach could be harder to maintain than using the `rawValue` data structure. This is especially true if support is added for handling more complex types (see #13).