@alex 2019-07-12
Ah, I hadn't thought about that. This could also be solved by sorting the
fields (though I shudder to think what would happen if you had overlapping
fields!)

On Fri, Jul 12, 2019 at 2:45 PM Ian Lopshire <notifications@github.com>
wrote:

> @alex <https://github.com/alex> @huydang284
> <https://github.com/huydang284>,
>
> The current implementation fails when the indices are not defined in order
> on the struct. For example:
>
> type people struct {
> 	ID        int     `fixed:"1,5"`
> 	LastName  string  `fixed:"16,25"`
> 	FirstName string  `fixed:"6,15"`
> 	Grade     float64 `fixed:"26,30"`
> }
>
> This could be solved by making sure values are written in order, but I
> think that approach could be harder to maintain than using the rawValue
> data structure. This is especially true if support is added for handling
> more complex types (see #13
> <https://github.com/ianlopshire/go-fixedwidth/issues/13>).
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/ianlopshire/go-fixedwidth/pull/12?email_source=notifications&email_token=AAAAGBDV7TVSSB3ZAOVCUSDP7DGOBA5CNFSM4H7CFAC2YY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGODZ2SD7Y#issuecomment-510992895>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAAAGBDGDNUFMY6ES22JC5LP7DGOBANCNFSM4H7CFACQ>
> .
>


-- 
All that is necessary for evil to succeed is for good people to do nothing.