@huydang284 2019-07-13
Oh, I see. But if we use `rawValue`, I think we still need sort fields so that we can get the right order.