@huydang284 2019-07-15
Hi @ianlopshire, I have a question.
Why don't we just use the fixed length (bytes or codepoints)? I think it is easier for implementing and better user experience also. When I have a struct with about 80 fields, I have to spend time to calculate the start and end position for each field, it's really hard to use.

My suggestion:
```
type Nested struct {
	F1 string `fixed:"10"`
	F2 struct {
		E1 string `fixed:"10"`
		E2 string `fixed:"10"`
	}
}
```