@ianlopshire 2019-07-15
@huydang284 feel free to open a separate issue to discuss the change. It is outside the scope of this PR.